requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
transformers
huggingface_hub
pydantic
tqdm
